#2 => 현황판 날씨 수정

+++ Ami/src/main/java/com/cnu/ami/device/server/dao/entity/ServerRegionIneterfaceVO.java
@@ -2,6 +2,6 @@ package com.cnu.ami.device.server.dao.entity;
 
 public interface ServerRegionIneterfaceVO {
 
-	long getRSEQ();
+	int getRSEQ();
 
 }
